azure-function-backend/requirements.txt
azure-functions
requests
azure-storage-blob
# Add other dependencies from your Flask backend as needed
